requirements.txt
absl-py==2.4.0
antlr4-python3-runtime==4.9.3
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dm-env==1.6
dm-tree==0.1.9
dm_control==1.0.36
etils==1.13.0
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
glfw==2.10.0
grpcio==1.78.0
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
labmaze==1.0.6
lxml==6.0.2
Markdown==3.10.2
MarkupSafe==2.1.5
matplotlib==3.10.8
mpmath==1.3.0
mujoco==3.4.0
networkx==3.6.1
numpy==2.3.5
omegaconf==2.3.0
opencv-python-headless==4.13.0.92
packaging==26.0
pandas==3.0.0
pillow==12.0.0
protobuf==6.33.5
PyOpenGL==3.1.10
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
setuptools==70.2.0
six==1.17.0
submitit==1.5.4
sympy==1.14.0
tb-nightly==2.21.0a20251023
tensorboard-data-server==0.7.2
termcolor==3.3.0
threadpoolctl==3.6.0
torch==2.10.0+cu128
torchaudio==2.10.0+cu128
torchvision==0.25.0+cu128
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
wrapt==2.1.1
zipp==3.23.0
